always compute CG functions in double precision

+++ Tutorials/gammaKK/gammaKKAmp/clebschGordan.h
@@ -1,7 +1,7 @@
 #if !defined(CLEBSCHGORDAN)
 #define CLEBSCHGORDAN
 
-float clebschGordan(int j1, int j2, int m1, int m2, int j, int m);
+double clebschGordan(int j1, int j2, int m1, int m2, int j, int m);
 
 double s3j(double j1, double j2, double j3, 
 		   double m1, double m2, double m3);
